Add files via upload
I have tried my best using the knowledge that I have. I am sure I can do more and better with the right set of knowledge and guidance.
Thank you for the opportunity. It was a lot of fun to work on something like this, and I discovered new things about myself and how much I like it. I am confident that I can put to a good use my aptitudes, creativeness, and abilities through programming.