Found a way to calculate the primes sum
I broke it down into steps, and the first step was to create a list of prime numbers between 2 set limits. I was looking for how to write a prime number in code language and I used loops in order to define all of them. After I got the list done I got blocked because I did not know how to add the things in the list, plus the list was including "1", that is not a prime number. I read more about Ruby and watched tutorials, and study it on codecademy, and I thought about adding the results into an array and after that just making a sum of the array elements. That worked just perfect and i used a "&&" operator in order to add to the arrays only the numbers higher than 1. I hope i did well :). Thank you